functions/requirements.txt

firebase-functions
firebase-admin
flask
tensorflow-cpu>=2.10
numpy
pillow
opencv-python-headless
werkzeug
